requirements.txt
Django==1.4.1
MySQL-python==1.2.3
South==0.7.6
django-articles==2.4.1
gunicorn==0.14.6
httplib2==0.7.4
oauth2==1.5.211
python-twitter==0.8.2
simplejson==2.6.1
wsgiref==0.1.2
